lots of improvements, mainly on separating rhz.Peering from application layer and some other things

+++ pkg/block/block.go
@@ -1,7 +1,7 @@
 package block
 
 type Header struct {
-	Index uint
+	Index uint64
 }
 
 type Block struct {
